requirements.txt
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tqdm>=4.66.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
